+++ src/components/Lunch/LunchCard/index.tsx
@@ -1,54 +1,8 @@
-import type { LunchMenuSummary } from '~/services/lunch';
+import { LunchCard as Card } from './LunchCard';
+import { LunchCardSkeleton } from './LunchCardSkeleton';
 
-import { css } from '@emotion/react';
-import React from 'react';
+export const LunchCard = Object.assign(Card, {
+  Skeleton: LunchCardSkeleton,
+});
 
-import { flex, palettes } from '~/styles/utils';
-
-import { LunchCardMenuDescription } from './LunchCardMenuDescription';
-import { LunchCardPollButton } from './LunchCardPollButton';
-
-export interface LunchCardProps {
-  polled?: boolean;
-  summary: LunchMenuSummary;
-  order: number;
-}
-
-const LunchCard = (props: LunchCardProps) => {
-  const { polled = false, summary, order } = props;
-  const { pollCount } = summary;
-
-  return (
-    <div css={selfCss}>
-      <LunchCardMenuDescription order={order} menu={summary} />
-
-      <LunchCardPollButton
-        css={{ width: likeButtonWidth }}
-        pollCount={pollCount}
-        polled={polled}
-        onPolledChange={() => {}}
-      />
-    </div>
-  );
-};
-export default LunchCard;
-
-const selfHeight = 182;
-const descriptionMinWidth = 200;
-const likeButtonWidth = 100;
-
-const selfMinWidth = descriptionMinWidth + likeButtonWidth;
-
-const selfCss = css(
-  {
-    position: 'relative',
-    backgroundColor: palettes.white,
-    borderRadius: 20,
-    overflow: 'hidden',
-    color: palettes.black,
-    minHeight: selfHeight,
-    height: '100%',
-    minWidth: selfMinWidth,
-  },
-  flex('stretch', '', 'row')
-);
+export type { LunchCardProps } from './LunchCard';
